Fix DeepSeek V3.2 FP8 weight loading, memory management, and runtime bugs
Weight loading:
- Fix postprocess_fp8_weights_batched: was passing 3D tensor to 2D-only
  deep_gemm.per_block_cast_to_fp8 (crash). Now iterates per-expert in-place
  matching vllm's requant_weight_ue8m0_inplace.
- Fix dequant precision: use float32 (not bfloat16) during UE8M0 requant.
- Use in-place .data.copy_() instead of nn.Parameter re-creation for FP8
  linear postprocessing.
- Add fastsafetensors GPU Direct Storage support for fast weight loading,
  matching vllm's fastsafetensors_weights_iterator.

Memory management:
- Share MoE FP8 workspace buffers (_ws13, _ws2, _out_buf) across all MoE
  layers. Without sharing, 58 layers x ~5GB each = ~285GB. Matches vllm's
  global WorkspaceManager.
- Fix BLOCK_SIZE mismatch: MLA uses block_size=64 but module-level
  BLOCK_SIZE was 256, causing block allocation/indexing mismatch
  (IndexError in prepare_mixed_batch). Now updates global BLOCK_SIZE
  when MLA cache is allocated.

Runtime fixes:
- Fix slot_mapping dtype: int32 -> int64 matching vllm.
- Fix sparse decode query absorption: absorb q_nope via W_UK_T before
  passing to FlashMLA sparse kernel which requires head_size_k=576.
- Fix _count_expert_tokens: replace graph-incompatible boolean indexing
  with Triton kernel matching vllm's _count_expert_num_tokens for CUDA
  graph compatibility.
- Fix block_tables=None guard in _forward_sparse_separate for initial
  prefill before blocks are allocated.
- Fix fp8_paged_mqa_logits: pass max_context_len positionally matching
  the C++ declaration (was using wrong keyword name max_model_len).

CUDA graph capture:
- Reduce capture sizes from 68 to 51 matching vllm's step-8/step-16
  pattern capped at 512.
- Single warmup at largest batch size instead of per-size warmup.
- Freeze GC during capture matching vllm.
- Total capture time reduced from ~594s to ~254s.

Made-with: Cursor

+++ tasks/baseline/L1/store_kvcache_fp8_mla.py
@@ -34,7 +34,7 @@ class StoreKVCacheFP8MLA(nn.Module):
         kv_c_normed: ``[N, 512]`` BF16 — compressed KV after layernorm.
         k_pe: ``[N, 1, 64]`` or ``[N, 64]`` BF16 — RoPE key component.
         kv_cache: ``[num_blocks, block_size, 656]`` uint8.
-        slot_mapping: ``[N]`` int32 — linear slot index per token (``-1`` skips).
+        slot_mapping: ``[N]`` int64 — linear slot index per token (``-1`` skips).
     """
 
     def __init__(self):
